requirements.txt
python-dotenv
langchain
langchain-community
pinecone>=3.0.0
sentence-transformers
python-docx
openai>=1.0.0
